+++ .travis.yml
@@ -9,7 +9,7 @@ env:
 addons:
   postgresql: "9.4"
 script:
-  - nosetests
+  - coverage run manage.py test
 notifications:
   slack: andela:
 after_success:
